refactor: apply structure on navigation
